+++ plugins/org.eclipse.mat.report/src/org/eclipse/mat/collect/HashMapIntObject.java
@@ -293,7 +293,7 @@ public final class HashMapIntObject<E> implements Serializable
     {
         capacity = PrimeFinder.findNextPrime(initialCapacity);
         step = PrimeFinder.findPrevPrime(initialCapacity);
-        limit = (int) ((float) capacity * 0.75);
+        limit = (int) (capacity * 0.75);
         clear();
         keys = new int[capacity];
         // This cast is ok as long as nobody assigns the field values to a field
