+++ plugins/org.eclipse.mat.report/src/org/eclipse/mat/collect/HashMapObjectLong.java
@@ -299,7 +299,7 @@ public final class HashMapObjectLong<E> implements Serializable
     {
         capacity = PrimeFinder.findNextPrime(initialCapacity);
         step = PrimeFinder.findPrevPrime(initialCapacity);
-        limit = (int) ((float) capacity * 0.75);
+        limit = (int) (capacity * 0.75);
         clear();
         keys = new Object[capacity];
         values = new long[capacity];
