+++ plugins/org.eclipse.mat.report/src/org/eclipse/mat/collect/SetLong.java
@@ -161,7 +161,7 @@ public final class SetLong implements Serializable
     {
         capacity = PrimeFinder.findNextPrime(initialCapacity);
         step = PrimeFinder.findPrevPrime(initialCapacity);
-        limit = (int) ((float) capacity * 0.75);
+        limit = (int) (capacity * 0.75);
         clear();
         keys = new long[capacity];
     }
